字典服务逻辑补充

+++ ticho-upms/ticho-upms-app/src/main/java/com/ticho/upms/infrastructure/entity/Dict.java
@@ -1,19 +1,20 @@
 package com.ticho.upms.infrastructure.entity;
 
-import java.io.Serializable;
-import java.time.LocalDateTime;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import com.fasterxml.jackson.annotation.JsonFormat;
+import com.baomidou.mybatisplus.annotation.FieldFill;
 import com.baomidou.mybatisplus.annotation.IdType;
+import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableId;
+import com.baomidou.mybatisplus.annotation.TableLogic;
 import com.baomidou.mybatisplus.annotation.TableName;
 import com.baomidou.mybatisplus.extension.activerecord.Model;
-import com.baomidou.mybatisplus.annotation.FieldFill;
-import com.baomidou.mybatisplus.annotation.TableField;
-import com.baomidou.mybatisplus.annotation.TableLogic;
+import com.fasterxml.jackson.annotation.JsonFormat;
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+import java.io.Serializable;
+import java.time.LocalDateTime;
 
 /**
  * 数据字典
@@ -31,11 +32,11 @@ public class Dict extends Model<Dict> implements Serializable {
     /** 主键编号; */
     @ApiModelProperty(value = "主键编号;", position = 10)
     @TableId(value = "id", type = IdType.ASSIGN_ID)
-    private Integer id;
+    private Long id;
 
     /** 父id */
     @ApiModelProperty(value = "父id", position = 20)
-    private Integer pid;
+    private Long pid;
 
     /** 字典类型id */
     @ApiModelProperty(value = "字典类型id", position = 30)
